Tested & gave up adjusting upward-shifting baseline
The code chunk lukeScratch1 contains code to load one test oyster W.O.5.1016 that had an upward trend in its baseline closed Hall value during August 2022. The code tries to correct that shift back down to a flat baseline, which works, but this introduces the situation that later calculations of percentage opening will underestimate the amount of valve opening when the oyster gets larger.